refactor: Split orbit-core command/search.rs (2,963 lines) into a searc… [ORB-00354]
Split orbit-core command/search.rs (2,963 lines) into a search/ module

Planned-By: codex
